+++ R/admiral_options.R
@@ -21,8 +21,7 @@
 #'
 #' @export
 #'
-#' @seealso [vars()], [set_admiral_options()], [derive_param_exist_flag()],
-#' [derive_param_first_event()], [derive_param_tte()], [derive_var_disposition_status()],
+#' @seealso [vars()], [set_admiral_options()], [derive_param_exist_flag()], [derive_param_tte()],
 #' [derive_var_dthcaus()], [derive_var_extreme_dtm()], [derive_vars_disposition_reason()],
 #' [derive_vars_period()], [create_period_dataset()]
 #'
@@ -81,8 +80,7 @@ get_admiral_option <- function(option) {
 #'
 #' @export
 #'
-#' @seealso [vars()], [get_admiral_option()], [derive_param_exist_flag()],
-#' [derive_param_first_event()], [derive_param_tte()], [derive_var_disposition_status()],
+#' @seealso [vars()], [get_admiral_option()], [derive_param_exist_flag()], [derive_param_tte()],
 #' [derive_var_dthcaus()], [derive_var_extreme_dtm()], [derive_vars_disposition_reason()],
 #' [derive_vars_period()], [create_period_dataset()]
 #'