+++ R/derive_merged.R
@@ -354,6 +354,7 @@ derive_vars_merged <- function(dataset,
 #' @author Stefan Bundfuss
 #'
 #' @keywords deprecated
+#' @family deprecated
 #'
 #' @export
 #'
@@ -372,45 +373,10 @@ derive_vars_merged_dt <- function(dataset,
                                   preserve = FALSE,
                                   check_type = "warning",
                                   duplicate_msg = NULL) {
-  assert_vars(by_vars)
-  dtc <- assert_symbol(enquo(dtc))
-  filter_add <- assert_filter_cond(enquo(filter_add), optional = TRUE)
-  assert_data_frame(dataset_add, required_vars = quo_c(by_vars, dtc))
-
-  deprecate_warn(
-    "0.8.0",
+  deprecate_stop(
+    "0.10.0",
     "derive_vars_merged_dt()",
     details = "Please use `derive_vars_dt()` and `derive_vars_merged()` instead."
-  )
-
-  old_vars <- names(dataset_add)
-  if (is.null(date_imputation)) {
-    highest_imputation <- "n"
-    date_imputation <- "first"
-  } else {
-    highest_imputation <- "M"
-  }
-  add_data <- filter_if(dataset_add, filter_add) %>%
-    derive_vars_dt(
-      new_vars_prefix = new_vars_prefix,
-      dtc = !!dtc,
-      highest_imputation = highest_imputation,
-      date_imputation = date_imputation,
-      flag_imputation = flag_imputation,
-      min_dates = min_dates,
-      max_dates = max_dates,
-      preserve = preserve
-    )
-  new_vars <- quos(!!!syms(setdiff(names(add_data), old_vars)))
-  derive_vars_merged(
-    dataset,
-    dataset_add = add_data,
-    by_vars = by_vars,
-    order = order,
-    new_vars = new_vars,
-    mode = mode,
-    check_type = check_type,
-    duplicate_msg = duplicate_msg
   )
 }
 
@@ -468,6 +434,7 @@ derive_vars_merged_dt <- function(dataset,
 #' @author Stefan Bundfuss
 #'
 #' @keywords deprecated
+#' @family deprecated
 #'
 #' @export
 #'
@@ -487,50 +454,11 @@ derive_vars_merged_dtm <- function(dataset,
                                    preserve = FALSE,
                                    check_type = "warning",
                                    duplicate_msg = NULL) {
-  dtc <- assert_symbol(enquo(dtc))
 
-  filter_add <- assert_filter_cond(enquo(filter_add), optional = TRUE)
-  assert_data_frame(dataset_add, required_vars = quo_c(by_vars, dtc))
-
-  deprecate_warn(
-    "0.8.0",
+  deprecate_stop(
+    "0.10.0",
     "derive_vars_merged_dtm()",
     details = "Please use `derive_vars_dtm()` and `derive_vars_merged()` instead."
-  )
-
-  old_vars <- names(dataset_add)
-  if (is.null(date_imputation)) {
-    highest_imputation <- "h"
-    date_imputation <- "first"
-  } else {
-    highest_imputation <- "M"
-  }
-  if (is.null(time_imputation)) {
-    highest_imputation <- "n"
-    time_imputation <- "first"
-  }
-  add_data <- filter_if(dataset_add, filter = filter_add) %>%
-    derive_vars_dtm(
-      new_vars_prefix = new_vars_prefix,
-      dtc = !!dtc,
-      highest_imputation = highest_imputation,
-      date_imputation = date_imputation,
-      time_imputation = time_imputation,
-      flag_imputation = flag_imputation,
-      min_dates = min_dates,
-      max_dates = max_dates,
-      preserve = preserve
-    )
-  new_vars <- quos(!!!syms(setdiff(names(add_data), old_vars)))
-  derive_vars_merged(
-    dataset,
-    dataset_add = add_data,
-    by_vars = by_vars,
-    order = order,
-    new_vars = new_vars,
-    mode = mode,
-    check_type = check_type,
-    duplicate_msg = duplicate_msg
   )
 }