+++ R/derive_param_computed.R
@@ -273,15 +273,5 @@ derive_derived_param <- function(dataset,
                                  filter = NULL,
                                  constant_by_vars = NULL,
                                  constant_parameters = NULL) {
-  deprecate_warn("0.8.0", "derive_derived_param()", "derive_param_computed()")
-  derive_param_computed(
-    dataset,
-    by_vars = by_vars,
-    parameters = parameters,
-    analysis_value = !!enquo(analysis_value),
-    set_values_to = set_values_to,
-    filter = !!enquo(filter),
-    constant_by_vars = constant_by_vars,
-    constant_parameters = constant_parameters
-  )
+  deprecate_stop("0.10.0", "derive_derived_param()", "derive_param_computed()")
 }